+++ nv/ex/parser/scanner_command_noremap.py
@@ -1,14 +1,14 @@
 from .tokens import TokenEof
-from .tokens_base import TOKEN_COMMAND_SMAP
+from .tokens_base import TOKEN_COMMAND_NOREMAP
 from .tokens_base import TokenOfCommand
 from NeoVintageous.nv import ex
 
 
-@ex.command('smap', 'sm')
-class TokenCommandSmap(TokenOfCommand):
+@ex.command('noremap', 'no')
+class TokenCommandNoremap(TokenOfCommand):
     def __init__(self, params, *args, **kwargs):
-        super().__init__(params, TOKEN_COMMAND_SMAP, 'smap', *args, **kwargs)
-        self.target_command = 'ex_smap'
+        super().__init__(params, TOKEN_COMMAND_NOREMAP, 'noremap', *args, **kwargs)
+        self.target_command = 'ex_noremap'
 
     @property
     def keys(self):
@@ -19,7 +19,7 @@ class TokenCommandSmap(TokenOfCommand):
         return self.params['command']
 
 
-def scan_command_smap(state):
+def scan_command_noremap(state):
     params = {
         'keys': None,
         'command': None,
@@ -30,4 +30,4 @@ def scan_command_smap(state):
     if m:
         params.update(m.groupdict())
 
-    return None, [TokenCommandSmap(params), TokenEof()]
+    return None, [TokenCommandNoremap(params), TokenEof()]