+++ nv/ex/parser/scanner_command_snoremap.py
@@ -1,14 +1,14 @@
 from .tokens import TokenEof
-from .tokens_base import TOKEN_COMMAND_NMAP
+from .tokens_base import TOKEN_COMMAND_SNOREMAP
 from .tokens_base import TokenOfCommand
 from NeoVintageous.nv import ex
 
 
-@ex.command('nmap', 'nm')
-class TokenCommandNmap(TokenOfCommand):
+@ex.command('snoremap', 'snor')
+class TokenCommandSnoremap(TokenOfCommand):
     def __init__(self, params, *args, **kwargs):
-        super().__init__(params, TOKEN_COMMAND_NMAP, 'nmap', *args, **kwargs)
-        self.target_command = 'ex_nmap'
+        super().__init__(params, TOKEN_COMMAND_SNOREMAP, 'snoremap', *args, **kwargs)
+        self.target_command = 'ex_snoremap'
 
     @property
     def keys(self):
@@ -19,7 +19,7 @@ class TokenCommandNmap(TokenOfCommand):
         return self.params['command']
 
 
-def scan_command_nmap(state):
+def scan_command_snoremap(state):
     params = {
         'keys': None,
         'command': None,
@@ -30,4 +30,4 @@ def scan_command_nmap(state):
     if m:
         params.update(m.groupdict())
 
-    return None, [TokenCommandNmap(params), TokenEof()]
+    return None, [TokenCommandSnoremap(params), TokenEof()]